Update to Ubuntu 14.04 + Apache2 + Php5 + MySQL 5.5.x; remover zip files and added wget download feature; magento 1.9.1.0 + sample used now

+++ vagrant/bootstrap.sh
@@ -1,34 +1,49 @@
 #!/usr/bin/env bash
 
-# Update
-# --------------------
+
+echo "=================================================="
+echo "Aloha! Now we will try to Install Ubuntu 14.04 LTS"
+echo "with Apache 2.4, PHP 5.6, MySQL 5.6(manual)"
+echo "and others dependencies needed for Magento 2."
+echo "Good luck :P"
+echo "=================================================="
+echo ""
+echo ""
+echo "=================================================="
+echo "SET LOCALES"
+echo "=================================================="
+
+export DEBIAN_FRONTEND=noninteractive
+
+export LANGUAGE=en_US.UTF-8
+export LANG=en_US.UTF-8
+export LC_TYPE=en_US.UTF-8
+export LC_ALL=en_US.UTF-8
+locale-gen en_US en_US.UTF-8
+dpkg-reconfigure locales
+
+
+echo "=================================================="
+echo "RUN UPDATE"
+echo "=================================================="
+
 apt-get update
+apt-get upgrade
 
 
-# Install Apache & PHP
-# --------------------
-apt-get install -y apache2
-apt-get install -y php5
-apt-get install -y libapache2-mod-php5
-apt-get install -y php5-mysql php5-curl php5-gd php5-intl php-pear php5-imap php5-mcrypt php5-ming php5-ps php5-pspell php5-recode php5-snmp php5-sqlite php5-tidy php5-xmlrpc php5-xsl php-apc
+echo "=================================================="
+echo "INSTALLING APACHE"
+echo "=================================================="
 
+apt-get -y install apache2
 
-# Delete default apache web dir and symlink mounted vagrant dir from host machine
-# --------------------
-rm -rf /var/www
-
-if [ ! -d "/vagrant/httpdocs" ]; then
-  mkdir /vagrant/httpdocs
+if ! [ -L /var/www ]; then
+  rm -rf /var/www
+  ln -fs /vagrant/httpdocs /var/www
 fi
 
-ln -fs /vagrant/httpdocs /var/www
-
-
-# Replace contents of default Apache vhost
-# --------------------
 VHOST=$(cat <<EOF
 <VirtualHost *:80>
-
   DocumentRoot "/var/www"
   ServerName simple-magento-vagrant.dev
 
@@ -37,19 +52,39 @@ VHOST=$(cat <<EOF
   </Directory>
 
   SetEnv MAGE_IS_DEVELOPER_MODE true
-
 </VirtualHost>
 EOF
 )
+echo "$VHOST" > /etc/apache2/sites-available/000-default.conf
 
-echo "$VHOST" > /etc/apache2/sites-enabled/000-default
+echo "ServerName localhost" | sudo tee /etc/apache2/conf-available/localhost.conf
 
+a2enconf localhost
 a2enmod rewrite
 service apache2 restart
 
 
-# Adminer
-# --------------------
+echo "=================================================="
+echo "INSTALLING PHP"
+echo "=================================================="
+
+apt-get -y update
+apt-get -y install php5 php5-mhash php5-mcrypt php5-curl php5-cli php5-mysql php5-gd php5-intl php5-xsl
+
+php5enmod mcrypt
+service apache2 reload
+
+
+echo "=================================================="
+echo "INSTALLING and CONFIGURE NTP"
+echo "=================================================="
+apt-get -y install ntp
+
+
+
+echo "=================================================="
+echo "INSTALLING ADMINER"
+echo "=================================================="
 if [ ! -d "/vagrant/httpdocs/adminer" ]; then
   echo "Adminer not found at /vagrant/httpdocs/adminer and will be installed..."
 
@@ -60,68 +95,58 @@ if [ ! -d "/vagrant/httpdocs/adminer" ]; then
 fi
 
 
-# Mysql
-# --------------------
-# Ignore the post install questions
-export DEBIAN_FRONTEND=noninteractive
-
-# Install MySQL quietly
+echo "=================================================="
+echo "INSTALLING MYSQL"
+echo "=================================================="
 apt-get -q -y install mysql-server-5.5
 
-mysql -u root -e "DROP DATABASE IF EXISTS magentodb"
-mysql -u root -e "CREATE DATABASE IF NOT EXISTS magentodb"
-mysql -u root -e "GRANT ALL PRIVILEGES ON magentodb.* TO 'magentouser'@'localhost' IDENTIFIED BY 'password'"
+echo "=================================================="
+echo "INSTALLING MYSQL MAGENTO DATABASE"
+echo "=================================================="
+
+mysql -u root -e "DROP DATABASE IF EXISTS magento"
+mysql -u root -e "CREATE DATABASE IF NOT EXISTS magento"
+mysql -u root -e "GRANT ALL PRIVILEGES ON magento.* TO 'magento'@'localhost' IDENTIFIED BY 'password'"
 mysql -u root -e "FLUSH PRIVILEGES"
 
 
-# Magento
-# --------------------
+echo "=================================================="
+echo "CLEANING..."
+echo "=================================================="
+apt-get -y autoremove
+apt-get -y autoclean
 
-#
-# Unpack magento
-# --------------------
-if [ -f "/vagrant/source/magento-1.8.1.0.tar.bz2" ]; then
-  echo "/vagrant/source/magento-1.8.1.0.tar.bz2 found. Start copy..."
-  tar xvf /vagrant/source/magento-1.8.1.0.tar.bz2 -C /vagrant/httpdocs/ --exclude='._*'
 
-  echo "moving files to /vagrant/httpdocs folder..."
-  mv /vagrant/httpdocs/magento/{*,.*} /vagrant/httpdocs
-
-  rm -r /vagrant/httpdocs/magento/
-
-  echo "Done."
-else
-  echo "/vagrant/source/magento-1.8.1.0.tar.bz2 not found."
+echo "=================================================="
+echo "DOWNLOAD MAGENTO SOURCE AND SAMPLE"
+echo "=================================================="
+echo "Start download Magento 1.9.1.0 and sample data save version..."
+mkdir /vagrant/source
+if [ magento-1.9.1.0.tar.gz does not exist ];
+then
+   wget -c http://www.magentocommerce.com/downloads/assets/1.9.1.0/magento-1.9.1.0.tar.gz
 fi
-
-
-#
-# Import DB
-# --------------------
-if [ -f "/vagrant/source/sql_magento_sample_data_1.6.1.0.sql" ]; then
-
-  echo "/vagrant/source/sql_magento_sample_data_1.6.1.0.sql found. Start import..."
-  mysql -u root magentodb < /vagrant/source/sql_magento_sample_data_1.6.1.0.sql
-  echo "Done. Run db update..."
-
-  #
-  # Update DB
-  # --------------------
-  mysql -u root -e "UPDATE magentodb.core_config_data SET value = 'http://simple-magento-vagrant.dev/' WHERE core_config_data.path = 'web/unsecure/base_url'"
-  mysql -u root -e "UPDATE magentodb.core_config_data SET value = 'http://simple-magento-vagrant.dev/' WHERE core_config_data.path = 'web/secure/base_url'"
-
-  echo "Update complete."
-else
-  echo "/vagrant/source/sql_magento_sample_data_1.6.1.0.sql not found."
+if [ magento-sample-data-1.9.1.0.tar.gz does not exist ];
+then
+   wget -c http://www.magentocommerce.com/downloads/assets/1.9.1.0/magento-sample-data-1.9.1.0.tar.gz
 fi
+echo "done."
+
+echo "Extract Magento and sample data to /vagrant/httpdocs ..."
+tar zxvf magento-1.9.1.0.tar.gz -C /vagrant/httpdocs --strip-components=1
+tar zxvf magento-sample-data-1.9.1.0.tar.gz -C /vagrant/httpdocs --strip-components=1
+echo "done."
+
+echo "Import Sample database..."
+mysql -u root magento < /vagrant/httpdocs/magento_sample_data_for_1.9.1.0.sql
+echo "done."
+
+echo "Update DB config with local domain name..."
+mysql -u root -e "UPDATE magento.core_config_data SET value = 'http://simple-magento-vagrant.dev/' WHERE core_config_data.path = 'web/unsecure/base_url'"
+mysql -u root -e "UPDATE magento.core_config_data SET value = 'http://simple-magento-vagrant.dev/' WHERE core_config_data.path = 'web/secure/base_url'"
+echo "done."
 
 
-# Import Media
-# --------------------
-if [ -f "/vagrant/source/media_magento_sample_data_for_1.6.1.0.tar.gz" ]; then
-  echo "/vagrant/source/media_magento_sample_data_for_1.6.1.0.tar.gz found. Start copy..."
-  tar xvzf /vagrant/source/media_magento_sample_data_for_1.6.1.0.tar.gz -C /vagrant/httpdocs/ --exclude='._*'
-  echo "Done."
-else
-  echo "/vagrant/source/media_magento_sample_data_for_1.6.1.0.tar.gz not found."
-fi
+echo "=================================================="
+echo "============= INSTALLATION COMPLETE =============="
+echo "=================================================="
